+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+0.5
+---
+
+ * [BC BREAK] The `host_url` configuration key for `ElevenLabs` has been renamed `endpoint`
+
 0.4
 ---